--- qiime/core/testing/markdown/method_bad_input_section.md
@@ -1,17 +1,0 @@
----
-name: Bad markdown
-description: "incorrectly formatted inputs"
-inputs:
-    - ints1:
-        - IntSequence1 | IntSequence2
-        - list
-        - Oops
-parameters:
-    - int1:
-        - Int
-        - int
-outputs:
-    - concatenated_ints:
-        - IntSequence1
-        - list
----
