--- qiime/core/testing/markdown/method_bad_outputs_section.md
@@ -1,17 +1,0 @@
----
-name: Bad markdown
-description: "incorrectly formatted outputs"
-inputs:
-    - ints1:
-        - IntSequence1 | IntSequence2
-        - list
-parameters:
-    - int1:
-        - Int
-        - int
-outputs:
-    - concatenated_ints:
-        - IntSequence1
-        - list
-        - Oops
----
