+++ qiime/plugin/tests/test_plugin.py
@@ -64,10 +64,9 @@ class TestPlugin(unittest.TestCase):
 
         self.assertIsInstance(actions, types.MappingProxyType)
         self.assertEqual(actions.keys(),
-                         {'merge_mappings', 'concatenate_ints',
-                          'concatenate_ints_markdown', 'split_ints',
-                          'split_ints_markdown', 'most_common_viz',
-                          'mapping_viz', 'identity_with_metadata',
+                         {'merge_mappings', 'concatenate_ints', 'split_ints',
+                          'most_common_viz', 'mapping_viz',
+                          'identity_with_metadata',
                           'identity_with_metadata_category'})
         for action in actions.values():
             self.assertIsInstance(action, qiime.sdk.Action)
@@ -82,9 +81,8 @@ class TestPlugin(unittest.TestCase):
         methods = self.plugin.methods
 
         self.assertEqual(methods.keys(),
-                         {'merge_mappings', 'concatenate_ints',
-                          'concatenate_ints_markdown', 'split_ints',
-                          'split_ints_markdown', 'identity_with_metadata',
+                         {'merge_mappings', 'concatenate_ints', 'split_ints',
+                          'identity_with_metadata',
                           'identity_with_metadata_category'})
         for method in methods.values():
             self.assertIsInstance(method, qiime.sdk.Action)
