+++ qiime/sdk/tests/test_method.py
@@ -13,77 +13,52 @@ import unittest
 import uuid
 
 import qiime.plugin
-from qiime.core.callable import Results
 from qiime.core.type import MethodSignature
-from qiime.sdk import Artifact, Action
+from qiime.sdk import Artifact, Method, Results
 
+from qiime.core.testing.method import (concatenate_ints, merge_mappings,
+                                       split_ints)
 from qiime.core.testing.type import IntSequence1, IntSequence2, Mapping
 from qiime.core.testing.util import get_dummy_plugin
 
 
-# TODO refactor these tests now that qiime.sdk.Action exists, and
-# Method/Visualizer classes have been removed. It probably makes sense to move
-# some of the tests here (and in Visualizer tests) into Action-specific test
-# cases (this will also remove some of the redundant tests betwen Method and
-# Visualizer).
+# TODO refactor these tests along with Visualizer tests to remove duplication.
 class TestMethod(unittest.TestCase):
     def setUp(self):
         self.plugin = get_dummy_plugin()
 
         self.concatenate_ints_sig = MethodSignature(
+            concatenate_ints,
             inputs={
-                'ints1': (IntSequence1 | IntSequence2, list),
-                'ints2': (IntSequence1, list),
-                'ints3': (IntSequence2, list)
+                'ints1': IntSequence1 | IntSequence2,
+                'ints2': IntSequence1,
+                'ints3': IntSequence2
             },
             parameters={
-                'int1': (qiime.plugin.Int, int),
-                'int2': (qiime.plugin.Int, int)
+                'int1': qiime.plugin.Int,
+                'int2': qiime.plugin.Int
             },
-            defaults={},
-            outputs=collections.OrderedDict([
-                ('concatenated_ints', (IntSequence1, list))
-            ])
+            outputs=[
+                ('concatenated_ints', IntSequence1)
+            ]
         )
 
         self.split_ints_sig = MethodSignature(
+            split_ints,
             inputs={
-                'ints': (IntSequence1, list)
+                'ints': IntSequence1
             },
             parameters={},
-            defaults={},
-            outputs=collections.OrderedDict([
-                ('left', (IntSequence1, list)),
-                ('right', (IntSequence1, list))
-            ])
+            outputs=[
+                ('left', IntSequence1),
+                ('right', IntSequence1)
+            ]
         )
 
     def test_private_constructor(self):
         with self.assertRaisesRegex(NotImplementedError,
-                                    'Action constructor.*private'):
-            Action()
-
-    def test_from_markdown_with_artifacts_and_parameters(self):
-        method = self.plugin.methods['concatenate_ints_markdown']
-
-        self.assertEqual(method.id, 'concatenate_ints_markdown')
-        self.assertEqual(method.signature, self.concatenate_ints_sig)
-        self.assertEqual(method.name, 'Concatenate integers')
-        self.assertTrue(
-            method.description.startswith('This method concatenates integers'))
-        self.assertTrue(
-            method.source.startswith('## Concatenate some integers'))
-
-    def test_from_markdown_with_multiple_outputs(self):
-        method = self.plugin.methods['split_ints_markdown']
-
-        self.assertEqual(method.id, 'split_ints_markdown')
-        self.assertEqual(method.signature, self.split_ints_sig)
-        self.assertEqual(method.name, 'Split sequence of integers in half')
-        self.assertTrue(
-            method.description.startswith('This method splits a sequence'))
-        self.assertTrue(
-            method.source.startswith('### Find midpoint'))
+                                    'Method constructor.*private'):
+            Method()
 
     def test_from_function_with_artifacts_and_parameters(self):
         method = self.plugin.methods['concatenate_ints']
@@ -102,15 +77,15 @@ class TestMethod(unittest.TestCase):
         self.assertEqual(method.id, 'split_ints')
 
         exp_sig = MethodSignature(
+            split_ints,
             inputs={
-                'ints': (IntSequence1, list)
+                'ints': IntSequence1
             },
             parameters={},
-            defaults={},
-            outputs=collections.OrderedDict([
-                ('left', (IntSequence1, list)),
-                ('right', (IntSequence1, list))
-            ])
+            outputs=[
+                ('left', IntSequence1),
+                ('right', IntSequence1)
+            ]
         )
         self.assertEqual(method.signature, exp_sig)
 
@@ -126,15 +101,15 @@ class TestMethod(unittest.TestCase):
         self.assertEqual(method.id, 'merge_mappings')
 
         exp_sig = MethodSignature(
+            merge_mappings,
             inputs={
-                'mapping1': (Mapping, dict),
-                'mapping2': (Mapping, dict)
+                'mapping1': Mapping,
+                'mapping2': Mapping
             },
             parameters={},
-            defaults={},
-            outputs=collections.OrderedDict([
-                ('merged_mapping', (Mapping, dict))
-            ])
+            outputs=[
+                ('merged_mapping', Mapping)
+            ]
         )
         self.assertEqual(method.signature, exp_sig)
 
@@ -146,27 +121,21 @@ class TestMethod(unittest.TestCase):
 
     def test_is_callable(self):
         self.assertTrue(callable(self.plugin.methods['concatenate_ints']))
-        self.assertTrue(
-            callable(self.plugin.methods['concatenate_ints_markdown']))
 
     def test_callable_properties(self):
         concatenate_ints = self.plugin.methods['concatenate_ints']
-        concatenate_ints_markdown = self.plugin.methods['concatenate_ints']
         merge_mappings = self.plugin.methods['merge_mappings']
 
-        for method in (concatenate_ints, concatenate_ints_markdown,
-                       merge_mappings):
+        for method in concatenate_ints, merge_mappings:
             self.assertEqual(method.__call__.__name__, '__call__')
             self.assertEqual(method.__call__.__annotations__, {})
             self.assertFalse(hasattr(method.__call__, '__wrapped__'))
 
     def test_async_properties(self):
         concatenate_ints = self.plugin.methods['concatenate_ints']
-        concatenate_ints_markdown = self.plugin.methods['concatenate_ints']
         merge_mappings = self.plugin.methods['merge_mappings']
 
-        for method in (concatenate_ints, concatenate_ints_markdown,
-                       merge_mappings):
+        for method in concatenate_ints, merge_mappings:
             self.assertEqual(method.async.__name__, 'async')
             self.assertEqual(method.async.__annotations__, {})
             self.assertFalse(hasattr(method.async, '__wrapped__'))
@@ -174,26 +143,24 @@ class TestMethod(unittest.TestCase):
     def test_callable_and_async_signature_with_artifacts_and_parameters(self):
         # Signature with input artifacts and parameters (i.e. primitives).
         concatenate_ints = self.plugin.methods['concatenate_ints']
-        concatenate_ints_markdown = self.plugin.methods['concatenate_ints']
 
-        for method in concatenate_ints, concatenate_ints_markdown:
-            for callable_attr in '__call__', 'async':
-                signature = inspect.Signature.from_callable(
-                    getattr(method, callable_attr))
-                parameters = list(signature.parameters.items())
+        for callable_attr in '__call__', 'async':
+            signature = inspect.Signature.from_callable(
+                getattr(concatenate_ints, callable_attr))
+            parameters = list(signature.parameters.items())
 
-                kind = inspect.Parameter.POSITIONAL_OR_KEYWORD
-                exp_parameters = [
-                    ('ints1', inspect.Parameter('ints1', kind)),
-                    ('ints2', inspect.Parameter('ints2', kind)),
-                    ('ints3', inspect.Parameter('ints3', kind)),
-                    ('int1', inspect.Parameter('int1', kind)),
-                    ('int2', inspect.Parameter('int2', kind))
-                ]
-                self.assertEqual(parameters, exp_parameters)
+            kind = inspect.Parameter.POSITIONAL_OR_KEYWORD
+            exp_parameters = [
+                ('ints1', inspect.Parameter('ints1', kind)),
+                ('ints2', inspect.Parameter('ints2', kind)),
+                ('ints3', inspect.Parameter('ints3', kind)),
+                ('int1', inspect.Parameter('int1', kind)),
+                ('int2', inspect.Parameter('int2', kind))
+            ]
+            self.assertEqual(parameters, exp_parameters)
 
-                self.assertEqual(signature.return_annotation,
-                                 inspect.Signature.empty)
+            self.assertEqual(signature.return_annotation,
+                             inspect.Signature.empty)
 
     def test_callable_and_async_signature_with_no_parameters(self):
         # Signature without parameters (i.e. primitives), only input artifacts.
@@ -216,77 +183,72 @@ class TestMethod(unittest.TestCase):
 
     def test_call_with_artifacts_and_parameters(self):
         concatenate_ints = self.plugin.methods['concatenate_ints']
-        concatenate_ints_markdown = \
-            self.plugin.methods['concatenate_ints_markdown']
 
         artifact1 = Artifact.import_data(IntSequence1, [0, 42, 43])
         artifact2 = Artifact.import_data(IntSequence2, [99, -22])
 
-        for method in concatenate_ints, concatenate_ints_markdown:
-            result = method(artifact1, artifact1, artifact2, 55, 1)
+        result = concatenate_ints(artifact1, artifact1, artifact2, 55, 1)
 
-            # Test properties of the `Results` object.
-            self.assertIsInstance(result, tuple)
-            self.assertIsInstance(result, Results)
-            self.assertEqual(len(result), 1)
-            self.assertEqual(result.concatenated_ints.view(list),
-                             [0, 42, 43, 0, 42, 43, 99, -22, 55, 1])
+        # Test properties of the `Results` object.
+        self.assertIsInstance(result, tuple)
+        self.assertIsInstance(result, Results)
+        self.assertEqual(len(result), 1)
+        self.assertEqual(result.concatenated_ints.view(list),
+                         [0, 42, 43, 0, 42, 43, 99, -22, 55, 1])
 
-            result = result[0]
+        result = result[0]
 
-            self.assertIsInstance(result, Artifact)
-            self.assertEqual(result.type, IntSequence1)
+        self.assertIsInstance(result, Artifact)
+        self.assertEqual(result.type, IntSequence1)
 
-            self.assertIsInstance(result.uuid, uuid.UUID)
+        self.assertIsInstance(result.uuid, uuid.UUID)
 
-            # Can retrieve multiple views of different type.
-            exp_list_view = [0, 42, 43, 0, 42, 43, 99, -22, 55, 1]
-            self.assertEqual(result.view(list), exp_list_view)
-            self.assertEqual(result.view(list), exp_list_view)
+        # Can retrieve multiple views of different type.
+        exp_list_view = [0, 42, 43, 0, 42, 43, 99, -22, 55, 1]
+        self.assertEqual(result.view(list), exp_list_view)
+        self.assertEqual(result.view(list), exp_list_view)
 
-            exp_counter_view = collections.Counter(
-                {0: 2, 42: 2, 43: 2, 99: 1, -22: 1, 55: 1, 1: 1})
-            self.assertEqual(result.view(collections.Counter),
-                             exp_counter_view)
-            self.assertEqual(result.view(collections.Counter),
-                             exp_counter_view)
+        exp_counter_view = collections.Counter(
+            {0: 2, 42: 2, 43: 2, 99: 1, -22: 1, 55: 1, 1: 1})
+        self.assertEqual(result.view(collections.Counter),
+                         exp_counter_view)
+        self.assertEqual(result.view(collections.Counter),
+                         exp_counter_view)
 
-            # Accepts IntSequence1 | IntSequence2
-            artifact3 = Artifact.import_data(IntSequence2, [10, 20])
-            result, = method(artifact3, artifact1, artifact2, 55, 1)
+        # Accepts IntSequence1 | IntSequence2
+        artifact3 = Artifact.import_data(IntSequence2, [10, 20])
+        result, = concatenate_ints(artifact3, artifact1, artifact2, 55, 1)
 
-            self.assertEqual(result.type, IntSequence1)
-            self.assertEqual(result.view(list),
-                             [10, 20, 0, 42, 43, 99, -22, 55, 1])
+        self.assertEqual(result.type, IntSequence1)
+        self.assertEqual(result.view(list),
+                         [10, 20, 0, 42, 43, 99, -22, 55, 1])
 
     def test_call_with_multiple_outputs(self):
         split_ints = self.plugin.methods['split_ints']
-        split_ints_markdown = self.plugin.methods['split_ints_markdown']
 
         artifact = Artifact.import_data(IntSequence1, [0, 42, -2, 43, 6])
 
-        for method in split_ints, split_ints_markdown:
-            result = method(artifact)
+        result = split_ints(artifact)
 
-            self.assertIsInstance(result, tuple)
-            self.assertEqual(len(result), 2)
+        self.assertIsInstance(result, tuple)
+        self.assertEqual(len(result), 2)
 
-            for output_artifact in result:
-                self.assertIsInstance(output_artifact, Artifact)
-                self.assertEqual(output_artifact.type, IntSequence1)
-                self.assertIsInstance(output_artifact.uuid, uuid.UUID)
+        for output_artifact in result:
+            self.assertIsInstance(output_artifact, Artifact)
+            self.assertEqual(output_artifact.type, IntSequence1)
+            self.assertIsInstance(output_artifact.uuid, uuid.UUID)
 
-            # Output artifacts have different UUIDs.
-            self.assertNotEqual(result[0].uuid, result[1].uuid)
+        # Output artifacts have different UUIDs.
+        self.assertNotEqual(result[0].uuid, result[1].uuid)
 
-            # Index lookup.
-            self.assertEqual(result[0].view(list), [0, 42])
-            self.assertEqual(result[1].view(list), [-2, 43, 6])
+        # Index lookup.
+        self.assertEqual(result[0].view(list), [0, 42])
+        self.assertEqual(result[1].view(list), [-2, 43, 6])
 
-            # Test properties of the `Results` object.
-            self.assertIsInstance(result, Results)
-            self.assertEqual(result.left.view(list), [0, 42])
-            self.assertEqual(result.right.view(list), [-2, 43, 6])
+        # Test properties of the `Results` object.
+        self.assertIsInstance(result, Results)
+        self.assertEqual(result.left.view(list), [0, 42])
+        self.assertEqual(result.right.view(list), [-2, 43, 6])
 
     def test_call_with_no_parameters(self):
         merge_mappings = self.plugin.methods['merge_mappings']
@@ -315,100 +277,80 @@ class TestMethod(unittest.TestCase):
 
     def test_async(self):
         concatenate_ints = self.plugin.methods['concatenate_ints']
-        concatenate_ints_markdown = \
-            self.plugin.methods['concatenate_ints_markdown']
 
         artifact1 = Artifact.import_data(IntSequence1, [0, 42, 43])
         artifact2 = Artifact.import_data(IntSequence2, [99, -22])
 
-        for method in concatenate_ints, concatenate_ints_markdown:
-            future = method.async(artifact1, artifact1, artifact2, 55, 1)
+        future = concatenate_ints.async(artifact1, artifact1, artifact2, 55, 1)
 
-            self.assertIsInstance(future, concurrent.futures.Future)
-            result = future.result()
+        self.assertIsInstance(future, concurrent.futures.Future)
+        result = future.result()
 
-            # Test properties of the `Results` object.
-            self.assertIsInstance(result, tuple)
-            self.assertIsInstance(result, Results)
-            self.assertEqual(len(result), 1)
-            self.assertEqual(result.concatenated_ints.view(list),
-                             [0, 42, 43, 0, 42, 43, 99, -22, 55, 1])
+        # Test properties of the `Results` object.
+        self.assertIsInstance(result, tuple)
+        self.assertIsInstance(result, Results)
+        self.assertEqual(len(result), 1)
+        self.assertEqual(result.concatenated_ints.view(list),
+                         [0, 42, 43, 0, 42, 43, 99, -22, 55, 1])
 
-            result = result[0]
+        result = result[0]
 
-            self.assertIsInstance(result, Artifact)
-            self.assertEqual(result.type, IntSequence1)
+        self.assertIsInstance(result, Artifact)
+        self.assertEqual(result.type, IntSequence1)
 
-            self.assertIsInstance(result.uuid, uuid.UUID)
+        self.assertIsInstance(result.uuid, uuid.UUID)
 
-            # Can retrieve multiple views of different type.
-            exp_list_view = [0, 42, 43, 0, 42, 43, 99, -22, 55, 1]
-            self.assertEqual(result.view(list), exp_list_view)
-            self.assertEqual(result.view(list), exp_list_view)
+        # Can retrieve multiple views of different type.
+        exp_list_view = [0, 42, 43, 0, 42, 43, 99, -22, 55, 1]
+        self.assertEqual(result.view(list), exp_list_view)
+        self.assertEqual(result.view(list), exp_list_view)
 
-            exp_counter_view = collections.Counter(
-                {0: 2, 42: 2, 43: 2, 99: 1, -22: 1, 55: 1, 1: 1})
-            self.assertEqual(result.view(collections.Counter),
-                             exp_counter_view)
-            self.assertEqual(result.view(collections.Counter),
-                             exp_counter_view)
+        exp_counter_view = collections.Counter(
+            {0: 2, 42: 2, 43: 2, 99: 1, -22: 1, 55: 1, 1: 1})
+        self.assertEqual(result.view(collections.Counter),
+                         exp_counter_view)
+        self.assertEqual(result.view(collections.Counter),
+                         exp_counter_view)
 
-            # Accepts IntSequence1 | IntSequence2
-            artifact3 = Artifact.import_data(IntSequence2, [10, 20])
-            future = method.async(artifact3, artifact1, artifact2, 55, 1)
-            result, = future.result()
+        # Accepts IntSequence1 | IntSequence2
+        artifact3 = Artifact.import_data(IntSequence2, [10, 20])
+        future = concatenate_ints.async(artifact3, artifact1, artifact2, 55, 1)
+        result, = future.result()
 
-            self.assertEqual(result.type, IntSequence1)
-            self.assertEqual(result.view(list),
-                             [10, 20, 0, 42, 43, 99, -22, 55, 1])
+        self.assertEqual(result.type, IntSequence1)
+        self.assertEqual(result.view(list),
+                         [10, 20, 0, 42, 43, 99, -22, 55, 1])
 
     def test_async_with_multiple_outputs(self):
         split_ints = self.plugin.methods['split_ints']
-        split_ints_markdown = self.plugin.methods['split_ints_markdown']
 
         artifact = Artifact.import_data(IntSequence1, [0, 42, -2, 43, 6])
 
-        for method in split_ints, split_ints_markdown:
-            future = method.async(artifact)
+        future = split_ints.async(artifact)
 
-            self.assertIsInstance(future, concurrent.futures.Future)
-            result = future.result()
+        self.assertIsInstance(future, concurrent.futures.Future)
+        result = future.result()
 
-            self.assertIsInstance(result, tuple)
-            self.assertEqual(len(result), 2)
+        self.assertIsInstance(result, tuple)
+        self.assertEqual(len(result), 2)
 
-            for output_artifact in result:
-                self.assertIsInstance(output_artifact, Artifact)
-                self.assertEqual(output_artifact.type, IntSequence1)
+        for output_artifact in result:
+            self.assertIsInstance(output_artifact, Artifact)
+            self.assertEqual(output_artifact.type, IntSequence1)
 
-                self.assertIsInstance(output_artifact.uuid, uuid.UUID)
+            self.assertIsInstance(output_artifact.uuid, uuid.UUID)
 
-            # Output artifacts have different UUIDs.
-            self.assertNotEqual(result[0].uuid, result[1].uuid)
+        # Output artifacts have different UUIDs.
+        self.assertNotEqual(result[0].uuid, result[1].uuid)
 
-            # Index lookup.
-            self.assertEqual(result[0].view(list), [0, 42])
-            self.assertEqual(result[1].view(list), [-2, 43, 6])
+        # Index lookup.
+        self.assertEqual(result[0].view(list), [0, 42])
+        self.assertEqual(result[1].view(list), [-2, 43, 6])
 
-            # Test properties of the `Results` object.
-            self.assertIsInstance(result, Results)
-            self.assertEqual(result.left.view(list), [0, 42])
-            self.assertEqual(result.right.view(list), [-2, 43, 6])
-
-    def test_markdown_input_section_validation(self):
-        with self.assertRaisesRegex(TypeError, 'input section'):
-            self.plugin.methods.register_markdown(
-                'markdown/method_bad_input_section.md')
-
-    def test_markdown_parameters_section_validation(self):
-        with self.assertRaisesRegex(TypeError, 'parameters section'):
-            self.plugin.methods.register_markdown(
-                'markdown/method_bad_parameters_section.md')
-
-    def test_markdown_outputs_section_validation(self):
-        with self.assertRaisesRegex(TypeError, 'outputs section'):
-            self.plugin.methods.register_markdown(
-                'markdown/method_bad_outputs_section.md')
+        # Test properties of the `Results` object.
+        self.assertIsInstance(result, Results)
+        self.assertEqual(result.left.view(list), [0, 42])
+        self.assertEqual(result.right.view(list), [-2, 43, 6])
 
 
 if __name__ == '__main__':
